Update red_pitaya_fads.sv
futher develop the file to make sure all variables are arrays that contain values for each channel, and add the respective outputs. I also implemented a kind of logger logic that averages the voltages in one multiplexer recording period and outputs it for each channel for external raw data tracking.